Refactor to limit exported interfaces. Comply to new naming convention.
