+++ src/types/APIResponseTypes.ts
@@ -1,6 +1,6 @@
-import { GenericSearchResultType } from './GenericSearchResultType';
+import { GenericSearchResultType, FacetsType } from './GenericSearchResultTypes';
 import { GenericRecordType } from './GenericRecordTypes';
-import { BroadcastRecordType } from './BroadcastRecordTypes';
+import { BroadcastRecordType } from './BroadcastRecordType';
 
 export interface APISearchResponseType {
 	data: {
@@ -13,14 +13,6 @@ export interface APISearchResponseType {
 		};
 		facet_counts: FacetsType;
 	};
-}
-
-export interface FacetsType {
-	facet_queries: Array<string>;
-	facet_fields: Array<string>;
-	facet_ranges: Array<string>;
-	facet_interval: Array<string>;
-	facet_heatmaps: Array<string>;
 }
 
 export interface APIRecordResponseType {
