+++ src/types/GenericSearchResultTypes.ts
@@ -1,4 +1,3 @@
-// Expand as we go
 export interface GenericSearchResultType {
 	access_billede_aftale: boolean;
 	access_blokeret: boolean;
@@ -30,4 +29,12 @@ export interface GenericSearchResultType {
 	type_of_resource: string[];
 	thumbnail: string;
 	_version: number;
+}
+
+export interface FacetsType {
+	facet_queries: Array<string>;
+	facet_fields: Array<string>;
+	facet_ranges: Array<string>;
+	facet_interval: Array<string>;
+	facet_heatmaps: Array<string>;
 }
